// src/api/generated/projects.d.ts
export type ActivityAt = string;
export type AssetId = number;
export type AssetType = "image" | "video";
export type Url = string;
export type CoverAssetId = number | null;
export type CreatedAt = string;
export type EpisodeCount = number;
export type Id = number;
export type Name1 = string;
export type Status = number;
export type UpdatedAt = string;
export type CoverAssetId1 = number | null;
export type CreatedAt1 = string;
export type CreatorId = number;
export type EpisodeNo = number;
export type Id1 = number;
export type Name2 = string;
export type ProjectId = number;
export type UpdatedAt1 = string;
export type Page = number;
export type PageSize = number;
export type Query = string;
export type Items = ProjectView[];
export type Page1 = number;
export type PageSize1 = number;
export type Total = number;
export type ProjectId1 = number;
export type Episodes = EpisodeView[];
// ...
export interface ProjectView {
  activity_at: ActivityAt;
  cover?: CoverView | null;
  cover_asset_id?: CoverAssetId;
  created_at: CreatedAt;
  episode_count?: EpisodeCount;
  id: Id;
  name: Name1;
  status: Status;
  updated_at: UpdatedAt;
  [k: string]: unknown;
}
export interface CoverView {
  asset_id: AssetId;
  asset_type: AssetType;
  url: Url;
  [k: string]: unknown;
}
export interface ProjectCreateResponse {
  default_episode: EpisodeView;
  project: ProjectView;
  [k: string]: unknown;
}
export interface EpisodeView {
  cover?: CoverView | null;
  cover_asset_id?: CoverAssetId1;
  created_at: CreatedAt1;
  creator_id: CreatorId;
  episode_no: EpisodeNo;
  id: Id1;
  name: Name2;
  project_id: ProjectId;
  updated_at: UpdatedAt1;
  [k: string]: unknown;
}
export interface ProjectListRequest {
  page?: Page;
  page_size?: PageSize;
  query?: Query;
  [k: string]: unknown;
}
export interface ProjectListResponse {
  items: Items;
  page?: Page1;
  page_size?: PageSize1;
  total?: Total;
  [k: string]: unknown;
}
export interface ProjectIdRequest {
  project_id: ProjectId1;
  [k: string]: unknown;
}
export interface ProjectDetailResponse {
  episodes: Episodes;
  project: ProjectView;
  [k: string]: unknown;
}